* updated version to 2.1 build 21
* updated system requirement to macOS 10.15 and later.
* fixed an auto launching issue when login.

+++ LoginItem Sample Launcher/AppDelegate.swift
@@ -10,19 +10,31 @@ import Cocoa
 
 @NSApplicationMain
 class AppDelegate: NSObject, NSApplicationDelegate {
-    
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         UserDefaults.shared.set(true, forKey: UserDefaults.Key.startFromLauncher.rawValue)
         
         let pathComponents = Bundle.main.bundleURL.pathComponents
         let mainRange = 0..<(pathComponents.count - 4)
         let mainPath = pathComponents[mainRange].joined(separator: "/")
-        try! NSWorkspace.shared.launchApplication(at: URL(fileURLWithPath: mainPath, isDirectory: false), options: [], configuration: [:])
-        NSApp.terminate(nil)
+        let url = URL(fileURLWithPath: mainPath, isDirectory: false)
+        let configuration = NSWorkspace.OpenConfiguration()
+        launchMainApp(url, with: configuration)
     }
 
     func applicationWillTerminate(_ aNotification: Notification) {
         // Insert code here to tear down your application
+    }
+    
+    private func launchMainApp(_ url:URL, with configuration:NSWorkspace.OpenConfiguration) {
+        NSWorkspace.shared.openApplication(at: url, configuration: configuration) { app, error in
+            if error == nil {
+                NSApp.terminate(nil)
+            } else {
+                DispatchQueue.main.asyncAfter(deadline: .now() + .seconds(2)) { [unowned self] in
+                    launchMainApp(url, with: configuration)
+                }
+            }
+        }
     }
 }
